fluxer-bot/requirements.txt
fluxer.py>=0.4.1
discord.py>=2.3.0
asyncpraw>=7.7.0
beautifulsoup4